+++ src/Compilers/CSharp/Test/Semantic/Semantics/PatternMatchingTests5.cs
@@ -1949,37 +1949,5 @@ _ = x is { Length.Error: > 0 };
                 Diagnostic(ErrorCode.ERR_NoSuchMember, "Error").WithArguments("int", "Error").WithLocation(7, 19)
                 );
         }
-
-        public class FlowAnalysisTests : FlowTestBase
-        {
-            [Fact]
-            public void RegionInIsPattern01()
-            {
-                var dataFlowAnalysisResults = CompileAndAnalyzeDataFlowExpression(@"
-class C
-{
-    static void M(object o)
-    {
-        _ = o switch
-        {
-            string { Length: 0 } s => /*<bind>*/s.ToString()/*</bind>*/,
-            _ = throw null
-        };
-    }
-}");
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.VariablesDeclared));
-                Assert.Equal("s", GetSymbolNamesJoined(dataFlowAnalysisResults.DataFlowsIn));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.DataFlowsOut));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.AlwaysAssigned));
-                Assert.Equal("s", GetSymbolNamesJoined(dataFlowAnalysisResults.ReadInside));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.WrittenInside));
-                Assert.Equal("o", GetSymbolNamesJoined(dataFlowAnalysisResults.ReadOutside));
-                Assert.Equal("o, s", GetSymbolNamesJoined(dataFlowAnalysisResults.WrittenOutside));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.Captured));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.CapturedInside));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.CapturedOutside));
-                Assert.Null(GetSymbolNamesJoined(dataFlowAnalysisResults.UnsafeAddressTaken));
-            }
-        }
     }
 }
